Try explicit instantiation in generated source files to speed up compilation
